refactor: mock server와 상관없는 okHttp 구현부를 ProductService 클래스로 분리하여 재사용 가능하도록 함
